Updated Drag and Drop + Test Pilot
I will be adding the snap update after my rest. I need to let my mind
chew on it for a bit. I finished the test pilot using the gear model,
modified the Drag and Drop functionality to expose the cartesian
coordinates, and added a level boundary for dragging the gears.