Master Scene Override
The Master Scene incorporates a bunch of new changes, additions,
deletions, and file moves. I added a new script that automatically
scales the GUI, created a working Toolbar, Fixed the Clipping
Backgrounds, added an Internal and Global section for scripts and
graphics shared between the games and menus, and added Scenes for menus.