Katherine's Heroic Commit
Fixed all Obstacle and Gear Colliders.
Updated Appearance of all Obstacle Models.
Gears now return to the Toolbox if they are not placed in a valid
location.
Programmatically Animated the Slide Gate.
Rebuilt and Fixed the Rotating Gate.
Scaled all Obstacles to reduce vision block.
Created new Tutorial-esque Puzzles to replace the old, broken puzzles.
Built a new recursive procedure to locate a powered Gear in a Gear and
Pylon chain.
Built Prefab Puzzle Templates to streamline level design.
Fixed the Layer Splash Graphic.
Fixed a Bug that prevented the obstacles from functioning.
Fixed a Bug that prevented gears from spawning in the Toolbox.
Fixed a lighting Bug that created random shadows on the Layers.
Updated the Gear Counter to scale with GUI
Rebuilt Gear and Obstacle Prefabs to snap to the layers, streamlining
level design.
Gears can now be placed in a Level as part of the Puzzle. They are
marked “Obstacle.”
Gears can be marked with “Forced Rotation”, thus creating new Power
Gears.
Rebuilt and fixed the “Layer Span Gear”
The Gear Spawning System can now be configured to force specific gears
to spawn in a certain order, in addition to randomly spawning gears.
Either may be disabled.
A hard cap can be placed on the number of Gears that spawn in the
puzzle.
A hard cap can be placed on the number of Gears available to select from
the Toolbox.
The relay system for Pylons has been updated to allow for one-way and
channel communications to one another directly.

Somehow, I managed this with no sleep or coffee...